requirements.txt
absl-py==0.10.0
addict==2.4.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
albumentations==0.5.2
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blessings==1.7
blinker==1.4
blis==0.7.4
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.1.1
catalogue==2.0.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.2
click==7.1.2
colorama==0.4.4
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
dlib @ https://pypi.python.org/packages/da/06/bd3e241c4eb0a662914b3b4875fc52dd176a9db0d4a2c915ac2ad8800e9e/dlib-19.7.0-cp36-cp36m-win_amd64.whl#md5=b7330a5b2d46420343fbed5df69e6a3f
Eel==0.14.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
face-detection==0.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==2.0.1
future==0.18.2
gast==0.3.3
gcloud==0.18.3
gevent==20.9.0
gevent-websocket==0.10.1
gitdb==4.0.5
GitPython==3.1.13
google-api-core==1.26.3
google-api-python-client==2.5.0
google-auth==1.22.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.8.4
greenlet==0.4.17
grpcio==1.33.1
h5py==2.10.0
hiredis==2.0.0
httplib2==0.19.1
idna==2.6
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
imantics==0.1.12
imgaug==0.4.0
immutables==0.15
importlib-metadata==1.7.0
importlib-resources==5.1.4
ipykernel==5.5.0
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jwcrypto==0.8
jws==0.1.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme2coco==0.1.2
lxml==4.6.3
Markdown==3.3.3
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib==3.3.3
mistune==0.8.4
moviepy==1.0.3
msgpack==1.0.2
multidict==5.1.0
munch==2.5.0
murmurhash==1.0.5
mxnet==1.7.0.post1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.2.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
opencensus==0.7.12
opencensus-context==0.1.2
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.4
pandocfilters==1.4.3
parso==0.8.1
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.0.1
pixellib==0.6.1
preshed==3.0.5
pretrainedmodels==0.7.4
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==1.0.18
protobuf==3.15.8
psutil==5.8.0
psycopg2-binary==2.8.6
py-spy==0.3.5
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.4
pydeck==0.6.1
pygame==2.0.0
Pygments==2.8.0
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
Pyrebase==3.0.27
Pyrebase4==4.4.3
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-jwt==2.0.1
python-twitter==3.5
python3-openid==3.2.0
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==3.13
pyzmq==22.0.3
ray==1.3.0
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.0
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
Shapely==1.7.1
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
smmap==3.0.5
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.0.1
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.2.19
sqlparse==0.4.1
srsly==2.4.1
streamlit==0.82.0
tabulate==0.8.9
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
thinc==8.0.3
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.3.2
toml==0.10.2
toolz==0.11.1
torch==1.8.1
torchtext==0.9.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
tune-sklearn==0.2.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.22
validators==0.18.2
virtualenv==20.4.7
wasabi==0.8.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xmljson==0.2.1
yarl==1.6.3
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
